Add geometry shading query (#19)
## What I'm adding
We need a way to communicate to the user visually the relationship
between the values for each geometry. I suggest interpolating the color
based on the feature's relative value compared to the max of the
selected scenario. We might well also want to allow a sub-selection
within the scenario. I believe this should be quite doable. Below is an
example of the "sh_energy" property from the first scenario. In any
case, we need to decide what will be the default for each scenario and
how / if we want users to have the ability to select between the various
combination of usage, category, and source.

<img width="959" alt="shading_features"
src="https://github.com/developmentseed/tecnico-energy-app/assets/31222040/858ab453-e31f-4951-9cee-0c65f33dae6d">