Add Github Actions to deploy separate preview branch
This feature creates a separate database and uses that for
the PR preview, allowing for end-users to preview the data
before deploying to production. A bot places a comment on
the PR to allow the users to easily find the PR preview.
To allow for this feature, we disabled PR previews on non-
production branches within Vercel.