Map legend should be based on max value, not total (ie sum) value (#92)
From Ricardo's email:

> one thing that would important to change when possible is the scale in
the dashboard. Now the scale/leged considers the Sum of the values as
the maximum and not the maximum value per se. with this the legend is
completely out of scale and scope.

This PR changes the legend so that the values are based on the _maximum_
value, not the _total_ value.